+++ types.js
@@ -677,10 +677,32 @@ export interface TONContracts {
 
     // Message processing
 
+    getMessageId(message: TONContractMessage): Promise<string>;
+
     sendMessage(
         params: TONContractMessage,
         parentSpan?: (Span | SpanContext)
     ): Promise<string>;
+
+
+    waitForTransaction(
+        message: TONContractMessage,
+        resultFields: string,
+        parentSpan?: (Span | SpanContext),
+        retryIndex?: number,
+    ): Promise<QTransaction>;
+
+    waitForDeployTransaction(
+        deployMessage: TONContractDeployMessage,
+        parentSpan?: (Span | SpanContext),
+        retryIndex?: number,
+    ): Promise<TONContractDeployResult>;
+
+    waitForRunTransaction(
+        runMessage: TONContractRunMessage,
+        parentSpan?: (Span | SpanContext),
+        retryIndex?: number,
+    ): Promise<TONContractRunResult>;
 
     processMessage(
         message: TONContractMessage,
@@ -764,11 +786,7 @@ type TONQueryAggregateFnType =
     | 'MIN'
     | 'MAX'
     | 'SUM'
-    | 'AVERAGE'
-    | 'STDDEV_POPULATION'
-    | 'STDDEV_SAMPLE'
-    | 'VARIANCE_POPULATION'
-    | 'VARIANCE_SAMPLE';
+    | 'AVERAGE';
 
 export type TONQueryAggregateField = {
     field: string,